refactor: PaymentMetadata to core/data, signing to core/auth
