+++ core/meterer/onchain_state.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"math"
-	"time"
 
 	"github.com/Layr-Labs/eigenda/core/eth"
 )
@@ -53,16 +52,9 @@ type OnDemandPayments struct {
 	Payments map[string]*OnDemandPayment
 }
 
-// TimeoutConfig contains the timeout configurations for the chain state
-type TimeoutConfig struct {
-	ChainReadTimeout  time.Duration
-	ChainStateTimeout time.Duration
-}
-
 // OnchainPaymentState is an interface for getting information about the current chain state for payments.
 type OnchainPaymentState struct {
-	tx            *eth.Transactor
-	TimeoutConfig TimeoutConfig
+	tx *eth.Transactor
 
 	ActiveReservations *ActiveReservations
 	OnDemandPayments   *OnDemandPayments
